test.py: pytest c++ facades should respect saving logs on success
BostFacade and UnitFacade saving the logs only when test failed,
ignoring the -s parameter that should allow save logs on success. This
PR adding checking this parameter.

+++ test/pylib/cpp/unit/unit_facade.py
@@ -51,5 +51,6 @@ class UnitTestFacade(CppTestFacade):
                 contents=msg
             )
             return [failure], ''
-        stdout_file_path.unlink(missing_ok=True)
+        if not self.save_log_on_success:
+            stdout_file_path.unlink(missing_ok=True)
         return None, ''
